feat: add configuration api route (#459)
* feat: add configuration api route


* feat: rename cache methods


---------
