+++ cmd/cache/remove.go
@@ -20,7 +20,6 @@ import (
 	"github.com/fatih/color"
 	"github.com/k8sgpt-ai/k8sgpt/pkg/cache"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
 
 // removeCmd represents the remove command
@@ -30,27 +29,11 @@ var removeCmd = &cobra.Command{
 	Long:  `This command allows you to remove the remote cache and use the default filecache.`,
 	Run: func(cmd *cobra.Command, args []string) {
 
-		// Remove the remote cache
-		var cacheInfo cache.CacheProvider
-		err := viper.UnmarshalKey("cache", &cacheInfo)
+		err := cache.RemoveRemoteCache(bucketname)
 		if err != nil {
 			color.Red("Error: %v", err)
 			os.Exit(1)
 		}
-		if cacheInfo.BucketName == "" {
-			color.Yellow("Error: no cache is configured")
-			os.Exit(1)
-		}
-		// Warn user this will delete the S3 bucket and prompt them to continue
-		color.Yellow("Warning: this will not delete the S3 bucket %s", cacheInfo.BucketName)
-		cacheInfo = cache.CacheProvider{}
-		viper.Set("cache", cacheInfo)
-		err = viper.WriteConfig()
-		if err != nil {
-			color.Red("Error: %v", err)
-			os.Exit(1)
-		}
-
 		color.Green("Successfully removed the remote cache")
 	},
 }
